Refactorized config extensions
