+++ ChustaSoft.Services.StaticData.AspNet/ChustaSoft.Services.StaticData.AspNet.csproj
@@ -23,9 +23,6 @@ https://github.com/Xelit3/StaticData/wiki</PackageReleaseNotes>
 
   <ItemGroup>
     <PackageReference Include="ChustaSoft.Common.AspNet" Version="1.0.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.ViewFeatures" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.2.4" />
   </ItemGroup>
 
   <ItemGroup>
