+++ ChustaSoft.Services.StaticData.IntegrationTest/TestHelpers/CurrencyTestHelper.cs
@@ -1,5 +1,5 @@
 ﻿using ChustaSoft.Services.StaticData.Base;
-using ChustaSoft.Services.StaticData.Helpers;
+using ChustaSoft.Services.StaticData.Configuration;
 using ChustaSoft.Services.StaticData.IntegrationTest.TestConstants;
 using ChustaSoft.Services.StaticData.Repositories;
 using ChustaSoft.Services.StaticData.Services;
