+++ ChustaSoft.Services.StaticData/ChustaSoft.Services.StaticData.csproj
@@ -1042,6 +1042,8 @@ https://github.com/Xelit3/StaticData/wiki</PackageReleaseNotes>
 
   <ItemGroup>
     <PackageReference Include="ChustaSoft.Common" Version="1.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
