+++ ChustaSoft.Services.StaticData/Configuration/StaticDataConfigurationBuilder.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ChustaSoft.Services.StaticData.Helpers
+namespace ChustaSoft.Services.StaticData.Configuration
 {
     public class StaticDataConfigurationBuilder : IStaticDataConfigurationBuilder
     {
